move fetching function for peminjaman to separate file
